% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{calculate_coverage}
\alias{calculate_coverage}
\title{Calculate Section Coverage}
\usage{
calculate_coverage(section, section_polygons, station, station_deployment,
  hourly_interval = 6)
}
\value{
A data frame with the columns Section, CoverageDateTime, Coverage and Receivers.
}
\description{
Calculates hourly proportional section coverage by hourly interval.
}
\examples{
calculate_coverage(qlexdatr::section, qlexdatr::section_polygons,
qlexdatr::station, qlexdatr::station_deployment)
}

