% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-detect-data.R
\name{check_detect_data}
\alias{check_detect_data}
\title{Check Lake Exploitation Detection Data}
\usage{
check_detect_data(data)
}
\arguments{
\item{data}{The detect_data object to check.}
}
\value{
A flag indicating whether the package data passes the checks.
}
\description{
Checks loaded lake exploitation data and returns a TRUE if passes all the tests.
Otherwise stops with an informative error.
}

