% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-lex-data.R
\name{check_lex_data}
\alias{check_lex_data}
\title{Check Lake Exploitation Data}
\usage{
check_lex_data(data, all = FALSE)
}
\arguments{
\item{data}{The lex_data object to check.}

\item{all}{A flag indicating whether to run deployment/detection checks.}
}
\value{
An invisible NULL if passes checks otherwise throws an informative error.
}
\description{
Checks loaded lake exploitation data and returns an invisible NULL if passes
all the tests.
Otherwise stops with an informative error.
}
\seealso{
input_lex_data
}
