% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ascii.R
\name{clean_ascii}
\alias{clean_ascii}
\title{Convert or force into ASCII.}
\usage{
clean_ascii(text)
}
\arguments{
\item{text}{Character string. String to be forced into ASCII.}
}
\value{
A clean ASCII character string.
}
\description{
Replaces non-ASCII charactes by ASCII equivalent and force ASCII encoding.
}
\author{
Nicolas Mangin
}
