% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_letters.R
\name{clean_letters}
\alias{clean_letters}
\title{Simplify string}
\usage{
clean_letters(text)
}
\arguments{
\item{text}{Character string. String to be simplified}
}
\value{
A simplified string.
}
\description{
Simplify the text to only lower case letters
}
\author{
Nicolas Mangin
}
