% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_paragraphs.R
\name{clean_paragraphs}
\alias{clean_paragraphs}
\title{Recompose paragraphs by identifying improper breaks.}
\usage{
clean_paragraphs(texts)
}
\arguments{
\item{texts}{Character vector. A vector of lines of text to be recomposed into paragraphs.}
}
\value{
A character vector in which paragraphs (ending with "." and with an empty line after) are recomposed.
}
\description{
Identify linebreaks which do not seem to match an end of paragraph and recompose the paragraph.
}
\author{
Nicolas Mangin
}
