% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_replace.R
\name{clean_replace}
\alias{clean_replace}
\title{Replace words in string.}
\usage{
clean_replace(text, replacements)
}
\arguments{
\item{text}{Character. A string in which patterns have to be replaced.}

\item{replacements}{Tibble. A tibble with two variables: "pattern" and "replacement".}
}
\value{
The initial string or vector in which replacements have been made.
}
\description{
Identify occurring patterns with a replacement and operate the replacement
}
\author{
Nicolas Mangin
}
