% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_spaces.R
\name{clean_spaces}
\alias{clean_spaces}
\title{Clean spaces}
\usage{
clean_spaces(text)
}
\arguments{
\item{text}{Character string. String to be cleaned.}
}
\value{
A clean string.
}
\description{
Clean spaces around punctuation, dashes, and parentheses.
}
\author{
Nicolas Mangin
}
