% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tags.R
\name{clean_tags}
\alias{clean_tags}
\title{Clean into pure ASCII text}
\usage{
clean_tags(text)
}
\arguments{
\item{text}{Character. String with HTML tags.}
}
\value{
String without HTML tags.
}
\description{
Remove HTML tags.
}
\author{
Nicolas Mangin
}
