% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_windows.R
\name{clean_windows}
\alias{clean_windows}
\title{Extract windows of words around a focal term.}
\usage{
clean_windows(text, term, window = c(3, 3))
}
\arguments{
\item{text}{Character string. String from which the windows should be extracted.}

\item{term}{Character string. Regex pattern around which the adjacent pattern has to be detected within the window.}

\item{window}{Numeric vector. Two values: one indicating the beginning of the window from the focal pattern, and one indicating the end.}
}
\value{
A list of windows around each occurrences of the focal term.
}
\description{
Extract windows of words around a regex pattern.
}
\author{
Nicolas Mangin
}
