% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_lex_data}
\alias{combine_lex_data}
\title{Combine Lake Exploitation Data}
\usage{
combine_lex_data(package)
}
\arguments{
\item{package}{A string of the data package.}
}
\description{
Combine Lake Exploitation Data
}

