% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-sections.R
\name{combine_sections_lex_data}
\alias{combine_sections_lex_data}
\title{Combine Sections Lex Data}
\usage{
combine_sections_lex_data(data, sections)
}
\arguments{
\item{data}{The lex_data object to combine the sections for.}

\item{sections}{A named list of the sections to combine.}
}
\value{
The modified lex
}
\description{
Combine Sections Lex Data
}
