% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{corp_info}
\alias{corp_info}
\title{Number of tokens in corpus (default: words)}
\usage{
corp_info(corpus, what = "word")
}
\arguments{
\item{corpus}{Corpus name}

\item{what}{Positional attribute defaulting to token (default "word")}
}
\description{
Wrapper function around rcqp::cqi_attribute_size
}

