% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bow.R
\name{create_bow}
\alias{create_bow}
\title{Create a bag of words}
\usage{
create_bow(
  text = NULL,
  basis = "lemma",
  keep_pos = c("NOUN", "PROPN", "ADJ", "VERB", "ADV")
)
}
\arguments{
\item{text}{Tibble or character string. Either syntactic relationships for one document, as returned by create_syntrel after grouping per document, or one document as a single simplified (see clean_letters) and ideally lemmatized string (see clean_replace and dat_en_lemmas).}

\item{basis}{Character string. Whether a bag of "word", a bag of "lemma", or a bag of "stem" should be returned (only for create_syntrel output).}

\item{keep_pos}{Character vector. Parts of speech which should be kept to build the bags of words (only for create_syntrel output)}
}
\value{
A tibble simplified into a bag of words.
}
\description{
Create bags of words from either an analysis of syntactic relationships or a character string.
}
\seealso{
\code{\link{create_syntrel}}

\code{\link{clean_letters}}

\code{\link{clean_replace}}

\code{\link{dat_en_lemmas}}
}
\author{
Nicolas Mangin
}
