% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dtm.R
\name{create_dtm}
\alias{create_dtm}
\title{Create a Document-to-Term Matrix}
\usage{
create_dtm(
  bow,
  min_term = 0,
  max_term = Inf,
  min_doc = 0,
  max_doc = Inf,
  nbterm = 1000,
  keep_terms = NULL,
  docvar = NULL
)
}
\arguments{
\item{bow}{Tibble. Output of the function eval_bow. Document ids must be in a variable called "document".}

\item{min_term}{Integer. Remove terms appearing less than this number of times.}

\item{max_term}{Integer. Remove terms appearing more than this number of times.}

\item{min_doc}{Integer, Remove terms appearing in less than this number of documents.}

\item{max_doc}{Integer, Remove terms appearing in more than this number of documents.}

\item{nbterm}{Integer. Select this number of terms based on tf-idf.}

\item{keep_terms}{Character vector. List of words which should be included even if they do not meet the other criteria.}

\item{docvar}{Tibble. Additional information about documents to be appended to the docvar of the dtm. Document ids must be in a variable called "document".}
}
\value{
A document to term matrix.
}
\description{
Transform bags of words into a document to term matrix after applying some filters.
}
\seealso{
\code{\link{eval_bow}}
}
\author{
Nicolas Mangin
}
