% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stm.R
\name{create_stm}
\alias{create_stm}
\title{Create a Structural Topic Model}
\usage{
create_stm(
  dtm,
  topic_nbr = 0,
  prevalence = NULL,
  content = NULL,
  init_type = "Spectral",
  seed = 1234,
  iterations = 50,
  tolerance = 0,
  keywords = NULL,
  baselab = "prob",
  frexweight = 0.6,
  min_gamma = 0.1,
  min_beta = 0.01,
  min_common = 1,
  verbose = TRUE
)
}
\arguments{
\item{dtm}{Document-to-Term matrix.}

\item{topic_nbr}{Numeric. Number of topics to model.}

\item{prevalence}{Formula. See the stm function for detailed instructions.}

\item{content}{Formula. See the stm function for detailed instructions.}

\item{init_type}{Character string. See the stm function for detailed instructions.}

\item{seed}{Numeric. Seed for initialization.}

\item{iterations}{Integer. Maximum number of iteration.}

\item{tolerance}{Numeric. See the stm function for detailed instructions.}

\item{keywords}{Tibble. Two variables: one for the "document" and one of the related "keywords".}

\item{baselab}{Character string. Whether labels should be based on "prob", "frex", "lift", or "score".}

\item{frexweight}{Numeric. The weight on discriminance in term selection.}

\item{min_gamma}{Numeric. Minimum strength of the relationship between document and topic.}

\item{min_beta}{Numeric. Minimum strength of the relationship between term and topic.}

\item{min_common}{Integer. Maximum number of topics allowed per document.}

\item{verbose}{Logical. Should messages be printed.}
}
\value{
A list with the "topic_model", the "topic_term", the "topic_label", and the "topic_document".
}
\description{
Create a topic model from a document-to-term matrix.
}
\seealso{
\code{\link[stm]{stm}}

\code{\link{create_dtm}}
}
\author{
Nicolas Mangin
}
