% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_syntrel.R
\name{create_syntrel}
\alias{create_syntrel}
\title{Create Syntactic Relationships}
\usage{
create_syntrel(text, model)
}
\arguments{
\item{text}{Character string. String to be parsed.}

\item{model}{Character string. Path from the working directory to the downloaded udpipe model.}
}
\value{
A tibble where each observation is a syntactic relationship between two words from the initial text.
}
\description{
Apply a udpipe model for part-of-speech tagging and dependency parsing to identify syntactic relationships.
}
\seealso{
\code{\link[udpipe]{udpipe}}
}
\author{
Nicolas Mangin
}
