% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_dictionaries.R
\docType{data}
\name{dat_dictionaries}
\alias{dat_dictionaries}
\title{Dictionaries for lexical analyses}
\format{
A tibble
}
\source{
\url{https://liwc.wpengine.com/}
}
\usage{
data("dat_dictionaries")
}
\description{
A tibble with patterns classified in various categories.
}
\author{
Nicolas Mangin
}
\keyword{LIWC}
\keyword{dictionary,}
