% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_en_lemmas.R
\docType{data}
\name{dat_en_lemmas}
\alias{dat_en_lemmas}
\title{English lemmatization}
\format{
A tiblle with 66,792 observations and 2 variables ("word" and "term")
}
\source{
\url{http://www.lexiconista.com/datasets/lemmatization/}
}
\usage{
data("dat_en_lemmas")
}
\description{
A table matching words with lemmas.
}
\author{
lexiconista, Nicolas Mangin
}
\keyword{lemmatization,}
\keyword{stem}
