% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_symbols.R
\docType{data}
\name{dat_symbols}
\alias{dat_symbols}
\title{List of symbols to be removed}
\format{
A character string with symbols.
}
\usage{
data("dat_symbols")
}
\description{
List of symbols to be removed
}
\author{
Nicolas Mangin
}
\keyword{cleaning}
\keyword{text}
