% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_toascii.R
\docType{data}
\name{dat_toascii}
\alias{dat_toascii}
\title{Map non ASCII to ASCII}
\format{
Tibble with special characters and their replacement.
}
\usage{
data("dat_toascii")
}
\description{
A table matching characters with their ASCII counterpart
}
\author{
Nicolas Mangin
}
\keyword{ASCII}
