% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_bow.R
\name{eval_bow}
\alias{eval_bow}
\title{Compute bag of words metrics}
\usage{
eval_bow(bow, document = "document", term = "term", count = "count")
}
\arguments{
\item{bow}{Tibble. Bag of words produced with create_bow.}

\item{document}{Character string. Name of the variable indicating the document id.}

\item{term}{Character string. Name of the variable indicating the term.}

\item{count}{Character string. Name of the variable indicating the frequence.}
}
\value{
A tibble 9 variables: document, term, count, doc_word_count (number of words in the document), term_count (frequency of the term), term_document_count (number of documents in which the term appears), tf, idf, tf_idf.
}
\description{
Compute basic metrics for a bag of words.
}
\seealso{
\code{\link{create_bow}}
}
\author{
Nicolas Mangin
}
