% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_dictionaries.R
\name{eval_dictionaries}
\alias{eval_dictionaries}
\title{Compute scores based on dictionaries' word counts}
\usage{
eval_dictionaries(bow, dictionary, score = "total")
}
\arguments{
\item{bow}{Tibble. Bag of words for one document with at least two variables: "term" and "count".}

\item{dictionary}{Tibble. Categorizations of words with at least three variables: "term", "category", and "value".}

\item{score}{Character string. Whether the function should return a "total" or a "percentage".}
}
\value{
A tibble with documents as rows, categories as variables, ad the required summary in cells
}
\description{
Count the occurrences of words associated with a dictionary.
}
\seealso{
\code{\link{create_bow}}
}
\author{
Nicolas Mangin
}
