% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_readability.R
\name{eval_readability}
\alias{eval_readability}
\title{Assess the readability of a text.}
\usage{
eval_readability(text)
}
\arguments{
\item{text}{Character string. Text.}
}
\value{
Various readability scores as one observation in a tibble.
}
\description{
Assess the readability of a text using various indicators.
}
\seealso{
\code{\link[koRpus]{readability}}
}
\author{
Nicolas Mangin
}
