% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_stm.R
\name{eval_stm}
\alias{eval_stm}
\title{Evaluate Structural Topic Models.}
\usage{
eval_stm(dtm = NA, topic_model = NA, keywords = NULL)
}
\arguments{
\item{dtm}{dtm. Document to term matrix used for Structural Topic Model.}

\item{topic_model}{stm. Structural Topic Model based on the document to term matrix.}

\item{keywords}{Tibble.}
}
\value{
A tibble with various metrics of tm quality for each possible combination of the topic number and seed.
}
\description{
Return various metrics to evaluate topic model performance for different topic numbers.
}
\seealso{
\code{\link{create_stm}}
}
\author{
Nicolas Mangin
}
