% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-detect-data.R
\name{filter_detect_data}
\alias{filter_detect_data}
\title{Filter Detect Data}
\usage{
filter_detect_data(
  data,
  capture = data$capture,
  section = data$section,
  alive_only = FALSE
)
}
\arguments{
\item{data}{The detect_data object.}

\item{capture}{A data frame of the capture data to use.}

\item{section}{A data frame of the section data to use.}

\item{alive_only}{A flag indicating wether to discard detections after last movement.}
}
\value{
A lex_data object.
}
\description{
Filter Detect Data
}
