% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-lex-data.R
\name{filter_lex_data}
\alias{filter_lex_data}
\title{Filter Lex Data}
\usage{
filter_lex_data(data, capture = data$capture, recapture = data$recapture,
  start_date = min(lubridate::date(capture$DateTimeCapture)),
  end_date = max(lubridate::date(capture$DateTimeTagExpire)))
}
\arguments{
\item{data}{The lex_data object.}

\item{capture}{A data frame of the capture data to use.}

\item{recapture}{A data frame of the recapture data to use.}

\item{start_date}{A date of the start.}

\item{end_date}{A date of the end.}
}
\value{
A lex_data object.
}
\description{
Filters lex data by capture and recapture.
Drops all recaptures and detections after first coded
harvest and ensures coded as tags removed.
}

