% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{find_word_id}
\alias{find_word_id}
\title{Find corpus id of the searched word}
\usage{
find_word_id(corpus, attr, what)
}
\arguments{
\item{corpus}{Corpus name}

\item{attr}{Positional attribute (lemma, word, pos etc.). Default: lemma}

\item{what}{Searched string}
}
\value{
Numeric identifier of the searched string in the selected corpus
}
\description{
Find corpus id of the searched word
}
\examples{
find_word('PATROLOGIA', 'lemma', 'tempus')
}

