% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{find_word_pos}
\alias{find_word_pos}
\title{Retrieve positions of the searched word}
\usage{
find_word_pos(corpus, attr, what.id)
}
\arguments{
\item{corpus}{Corpus name}

\item{attr}{Positional attribute (lemma, word, pos etc.). Default: lemma}

\item{what.id}{Numeric id to be searched in the corpus (retrieved with find_word_id)}
}
\value{
Vector containing positions of the searched string in the selected corpus
}
\description{
Retrieve positions of the searched word
}
\examples{
find_word('PATROLOGIA', 'lemma', 232)
}

