% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orig.R
\name{forms_stats}
\alias{forms_stats}
\title{Basic word form statistics}
\usage{
forms_stats(string, corp_pos, corp_size)
}
\arguments{
\item{string}{String of form CORPUS_NAME.attribute}

\item{corp_pos}{Vector of node word positions}

\item{corp_size}{Number of tokens in corpus}
}
\description{
Calculates wordform distribution for selected node
}
\examples{
forms_stats("PATROLOGIA.word", c(123,236,649), 102000000)
}

