% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-difftime.R
\name{get_difftime}
\alias{get_difftime}
\title{Get Difftime}
\usage{
get_difftime(object)
}
\arguments{
\item{object}{The object to get the difftime for.}
}
\value{
A difftime object
}
\description{
Get Difftime
}

