% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth_no}
\alias{growth_no}
\title{No Growth}
\usage{
growth_no(length, years)
}
\arguments{
\item{length}{A count indicating the length at capture.}

\item{years}{A number indicating the number of years since capture.}
}
\description{
Function that simply returns the input length, ie allows no growth.
}
\examples{
growth_no(100L, 0)
growth_no(100L, 1)
growth_no(100L, -1)
}

