% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth_vb}
\alias{growth_vb}
\title{Von Bertalanffy Growth}
\usage{
growth_vb(length, years, linf = 1000, k = 0.19)
}
\arguments{
\item{length}{A count indicating the length at capture.}

\item{years}{A number indicating the number of years since capture.}

\item{linf}{Length at infinity for Von Bertalanffy Growth Curve.}

\item{k}{Growth coefficient for Von Bertalanffy Growth Curve.}
}
\description{
Function that calculates now length based on Von Bertalanffy growth.
}
\examples{
growth_vb(100L, 0)
growth_vb(100L, 1)
growth_vb(100L, -1)
growth_vb(2000L, 1)
}
