% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-lex-data.R
\name{input_lex_data}
\alias{input_lex_data}
\title{Inputs Lake Exploitation Data}
\usage{
input_lex_data(package = "klexdatr")
}
\arguments{
\item{package}{A string of the data package.}
}
\description{
Inputs lake exploitation
data from an R data package structured like \code{klexdatr}.
}
