% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.analysis_data}
\alias{is.analysis_data}
\title{Is Analysis Data}
\usage{
is.analysis_data(x)
}
\arguments{
\item{x}{The object to test}
}
\value{
A flag indicating whether x is analysis data.
}
\description{
Is Analysis Data
}
