% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.detect_data}
\alias{is.detect_data}
\title{Is Detect Data}
\usage{
is.detect_data(x)
}
\arguments{
\item{x}{The object to test}
}
\value{
A flag indicating whether x is detect data.
}
\description{
Is Detect Data
}
