% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last-section-data.R
\name{last_section_data}
\alias{last_section_data}
\title{Last Section Data}
\usage{
last_section_data(data)
}
\arguments{
\item{data}{The detect_data object to merge.}
}
\value{
A tbl data frame.
}
\description{
Calculates number of fish last detected at each section.
}

