% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexr-package.R
\docType{package}
\name{lexr-package}
\alias{lexr}
\alias{lexr-package}
\title{lexr: Lake Exploitation Study Data Plotting and Analysis}
\description{
An R package to plot and analyse lake exploitation data.
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca}

Authors:
\itemize{
  \item Gary Pavan
  \item Bernhard Konrad
}

}
\keyword{internal}
