% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{lexr}
\alias{lexr}
\alias{lexr-package}
\title{Lake Exploitation Study Data Plotting and Analysis}
\description{
Lake Exploitation Study Data Plotting and Analysis
}
\examples{
library(qlexdatr)
library(lexr)
data <- input_lex_data("qlexdatr")
check_lex_data(data)
print(make_detect_data(data))
# print(data)
# plot(data)
}

