% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_lex_data}
\alias{load_lex_data}
\title{Load Lake Exploitation Data}
\usage{
load_lex_data(package)
}
\arguments{
\item{package}{A string of the data package.}
}
\description{
Load Lake Exploitation Data
}

