% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-analysis-data.R
\name{make_analysis_data}
\alias{make_analysis_data}
\title{Make Analysis Data}
\usage{
make_analysis_data(data, capture = data$capture, section = data$section)
}
\arguments{
\item{data}{A detect_data object to use.}

\item{capture}{A data frame of the capture data to use.}

\item{section}{A data frame of the section data to use.}
}
\value{
A detect_data object.
}
\description{
Makes analysis_data object from a detect_data object.
}

