% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-detect-data.R
\name{make_detect_data}
\alias{make_detect_data}
\title{Make Detect Data}
\usage{
make_detect_data(data, capture = data$capture,
  start_date = min(as.Date(capture$DateTimeCapture)),
  end_date = min(Sys.Date(), max(as.Date(capture$DateTimeTagExpire))),
  hourly_interval = 6L)
}
\arguments{
\item{data}{The lex_data object to check.}

\item{capture}{A data frame of the capture data to use.}

\item{hourly_interval}{A count indicating the hourly interval.}
}
\value{
A detect_data object.
}
\description{
Makes detect_data object from a lex_data object.
}

