% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-detect-data.R
\name{merge_detect_data}
\alias{merge_detect_data}
\title{Merge Detect Data}
\usage{
merge_detect_data(data)
}
\arguments{
\item{data}{The detect_data object to merge.}
}
\value{
A tbl data frame.
}
\description{
Merges Detect data into a single data frame which is useful for
calculating habitat use.
}
