% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orig.R
\name{node_stats}
\alias{node_stats}
\title{Basic node word statistics}
\usage{
node_stats(corp_pos, corp_size)
}
\arguments{
\item{corp_pos}{Vector of node word positions}

\item{corp_size}{Number of tokens in corpus}
}
\description{
Calculates node statistics basing on corpus position vector
}

