% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-analysis-data.R
\name{plot.analysis_data}
\alias{plot.analysis_data}
\title{Plot Analysis Data}
\usage{
\method{plot}{analysis_data}(x, all = FALSE, ...)
}
\arguments{
\item{x}{The analysis_data object to plot.}

\item{all}{A flag indicating whether to produce all plots.}

\item{...}{unused.}
}
\description{
Plot Analysis Data
}

