% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-detect-data.R
\name{plot.detect_data}
\alias{plot.detect_data}
\title{Plot Detect Data}
\usage{
\method{plot}{detect_data}(x, all = FALSE, ...)
}
\arguments{
\item{x}{The detect_data object to plot.}

\item{all}{A flag indicating whether to produce all plots.}

\item{...}{unused.}
}
\description{
Plot Detect Data
}

