% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-lex-data.R
\name{plot.lex_data}
\alias{plot.lex_data}
\title{Plot Lex Data}
\usage{
\method{plot}{lex_data}(x, all = FALSE, ...)
}
\arguments{
\item{x}{The lex_data object to plot.}

\item{all}{A flag indicating whether to produce all plots.}

\item{...}{unused.}
}
\description{
Plot Lex Data
}

