% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-capture.R
\name{plot_capture}
\alias{plot_capture}
\title{Plot Captures}
\usage{
plot_capture(data)
}
\arguments{
\item{data}{The detect_data object to plot.}
}
\description{
Plots captures for a \code{detect_data} object.
}
