% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data-dir.R
\name{plot_data_dir}
\alias{plot_data_dir}
\title{Plot Data to Directory}
\usage{
plot_data_dir(data, all = FALSE)
}
\arguments{
\item{data}{The data object to plot}

\item{all}{A flag indicating whether to plot all the data.}
}
\description{
Plot Data to Directory
}

