% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-detect-data.R
\name{plot_detect_coverage}
\alias{plot_detect_coverage}
\title{Plot Coverage}
\usage{
plot_detect_coverage(data)
}
\arguments{
\item{data}{The \code{detect_data} object to plot.}
}
\description{
Plots the color-coded percent receiver coverage for a \code{detect_data} object
produced by the \code{\link{make_detect_data}} function.
}
