% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-detect-data.R
\name{plot_detect_distance}
\alias{plot_detect_distance}
\title{Plot Detection Distance}
\usage{
plot_detect_distance(data)
}
\arguments{
\item{data}{The \code{detect_data} object to plot.}
}
\description{
Plots the color-coded distances between sections for a \code{detect_data} object
produced by the \code{\link{make_detect_data}} function.
}

