% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-detect-data.R
\name{plot_detect_overview}
\alias{plot_detect_overview}
\title{Plot Detection Overview}
\usage{
plot_detect_overview(data)
}
\arguments{
\item{data}{The \code{detect_data} object to plot.}
}
\description{
Plots the color-coded detection overview plot for a \code{detect_data} object
produced by the \code{\link{make_detect_data}} function.
}

