% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lex_data}
\alias{plot_lex_data}
\title{Plot Lex Data}
\usage{
plot_lex_data(package, file = paste0(package, ".pdf"))
}
\arguments{
\item{package}{A string of the data package.}

\item{file}{A string of the pdf file name.}
}
\description{
Plots lake exploitation package data to a pdf.
}
\examples{
\dontrun{
# creates qlexdatr.pdf in the working directory
plot_lex_data("qlexdatr")
}
}

