% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-capture.R
\name{plot_recapture}
\alias{plot_recapture}
\title{Plot Recaptures}
\usage{
plot_recapture(data)
}
\arguments{
\item{data}{The detect_data object to plot.}
}
\description{
Plots recaptures for a \code{detect_data} object.
}

