% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_section}
\alias{plot_section}
\title{Plot Section}
\usage{
plot_section(section, section_polygons = NULL)
}
\arguments{
\item{section}{A data frame of the section data.}

\item{section_polygons}{A xx.}
}
\value{
A ggplot2 object.
}
\description{
Plot Section
}
\examples{
plot_section(qlexdatr::section, qlexdatr::section_polygons)
}

