% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_station_deployment}
\alias{plot_station_deployment}
\title{Plot Station}
\usage{
plot_station_deployment(station_deployment, station)
}
\arguments{
\item{station_deployment}{A data frame of the receiver deployment data.}

\item{station}{A data frame of the station data.}
}
\value{
A ggplot2 object.
}
\description{
Plot Station
}
\examples{
plot_station_deployment(qlexdatr::station_deployment, qlexdatr::station)
}

