% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-detect-data.R
\name{plot_use_detect}
\alias{plot_use_detect}
\title{Plot Use}
\usage{
plot_use_detect(data, alive_only = TRUE)
}
\arguments{
\item{data}{The detect_data object to merge.}

\item{alive_only}{A flag indicating wether to discard detections after last movement.}
}
\description{
Plots habitat use for a \code{detect_data} object.
}
