% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_topic.R
\name{predict_topic}
\alias{predict_topic}
\title{Apply a STM to a new corpus}
\usage{
predict_topic(
  bow,
  topic_model,
  keywords = NULL,
  min_gamma = 0.1,
  min_common = 1
)
}
\arguments{
\item{bow}{Tibble. Bag of words as returned by the function create_bow}

\item{topic_model}{List. Structural topic model as returned by the function create_stm.}

\item{keywords}{Tibble. One variable specifying the "document", one variable containing the "keywords" as a character vector.}

\item{min_gamma}{Numeric. Minimum strength of the relationship between topic and document.}

\item{min_common}{Integer. Maximum number of topics allowed per document.}
}
\value{
A tibble linking topics and documents for a new corpus.
}
\description{
Apply a topic model to a new corpus of documents.
}
\seealso{
\code{\link{create_bow}}

\code{\link{create_stm}}
}
\author{
Nicolas Mangin
}
