% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lexical_dispersion.R
\name{print.lexical_dispersion}
\alias{print.lexical_dispersion}
\title{Prints a lexical_dispersion Object}
\usage{
\method{print}{lexical_dispersion}(x, plot = TRUE, ...)
}
\arguments{
\item{x}{lexical_dispersion object.}

\item{plot}{logical.  If \code{TRUE} the plot will automatically plot.
The user may wish to set to \code{FALSE} for use in knitr, sweave, etc.
to add additional plot layers.}

\item{\ldots}{Ignored.}
}
\description{
Prints a lexical_dispersion object.
}

