% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SpatialPolygonsDataFrame}
\alias{print.SpatialPolygonsDataFrame}
\title{Print Spatial Polygons Data Frame}
\usage{
\method{print}{SpatialPolygonsDataFrame}(x, ...)
}
\arguments{
\item{x}{The SpatialPolygonsDataFrame object to plot.}

\item{...}{unused.}
}
\description{
Print Spatial Polygons Data Frame
}
