% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_lex_data}
\alias{process_lex_data}
\title{Process Data}
\usage{
process_lex_data(package, hourly_interval = 6)
}
\arguments{
\item{hourly_interval}{A count indicating the hourly interval.}
}
\value{
A list of the data frames and transition matrix.
}
\description{
Process Data
}
\examples{
\dontrun{
process_lex_data("qlexdatr")
}
}

