% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-detect-data.R
\name{section_polygon}
\alias{section_polygon}
\title{Section Polygon}
\usage{
section_polygon(section)
}
\arguments{
\item{section}{The SpatialPolygonsDataFrame object to convert}
}
\value{
A data frame.
}
\description{
Converts section SpatialPolygonsDataFrame into data frame for plotting.
}

