% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spawning.R
\name{spawning_no}
\alias{spawning_no}
\title{No Spawning}
\usage{
spawning_no(detection, capture, period)
}
\arguments{
\item{detection}{A data.frame of the detection data for the capture.}

\item{years}{A data.frame of the period data.}
}
\description{
Function that simply returns FALSE for every time period.
}
\details{
detection has the columns Section, DateTime and Interval while
period has the columns Period and Interval for all period.
Any replacement function has to use this information to return a logical
vector indicating whether the individual was spawing in each period.
}

