% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_regex.R
\name{test_regex}
\alias{test_regex}
\title{Test Regular Expressions}
\usage{
test_regex(dictionary, texts)
}
\arguments{
\item{dictionary}{Tibble. Dictionary with at least a column "pattern".}

\item{texts}{Character vector. List of texts.}
}
\value{
A new dictionary
}
\description{
Gadget to test and adjust regular expressions in a dictionary.
}
\author{
Nicolas Mangin
}
