% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tidy_section_polygons}
\alias{tidy_section_polygons}
\title{Tidy Section Polygons}
\usage{
tidy_section_polygons(section_polygons)
}
\arguments{
\item{section_polygons}{A spatial polygons data frame of the sections
or NULL.}
}
\description{
Tidy Section Polygons
}

