% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-detect-data.R
\name{use_detect_data}
\alias{use_detect_data}
\title{Use Detect Data}
\usage{
use_detect_data(data, alive_only = TRUE)
}
\arguments{
\item{data}{The detect_data object to merge.}

\item{alive_only}{A flag indicating wether to discard detections after last movement.}
}
\value{
A tbl data frame.
}
\description{
Calculates Percent Use from Detect data.
}

