% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-analysis-data.R
\name{check_analysis_data}
\alias{check_analysis_data}
\title{Check Lake Exploitation Analysis Data}
\usage{
check_analysis_data(data)
}
\arguments{
\item{data}{The \code{analysis_data} object to check.}
}
\description{
Checks lake exploitation analysis data and returns an invisible copy of the data.
Otherwise stops with an informative error.
}
