% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-detect-data.R
\name{check_detect_data}
\alias{check_detect_data}
\title{Check Lake Exploitation Detection Data}
\usage{
check_detect_data(data)
}
\arguments{
\item{data}{The \code{detect_data} object to check.}
}
\description{
Checks lake exploitation data and returns an invisible copy of the data.
Otherwise stops with an informative error.
}
