% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_lex_data}
\alias{check_lex_data}
\title{Check Lake Exploitation Data}
\usage{
check_lex_data(data)
}
\arguments{
\item{data}{The lex_data object to check.}
}
\value{
A flag indicating whether the package data passes the checks.
}
\description{
Checks loaded lake exploitation data and returns a TRUE if passes all the tests.
Otherwise stops with an informative error.
}

