% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_lex_data}
\alias{check_lex_data}
\title{Check Lake Exploitation Data}
\usage{
check_lex_data(package)
}
\arguments{
\item{package}{A string of the data package.}
}
\value{
A flag indicating whether the package data passes the checks.
}
\description{
Checks lake exploitation data and returns a TRUE if passes all the tests.
Otherwise stops with an informative error.
}

