% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_windows.R
\name{clean_windows}
\alias{clean_windows}
\title{Extract windows of words around a focal pattern.}
\usage{
clean_windows(text, pattern, before = 3, after = 3)
}
\arguments{
\item{text}{Character string. String from which the windows should be extracted.}

\item{pattern}{Character string. Regex pattern around which the adjacent pattern has to be detected within the window.}

\item{before}{Integer. Number of words to extract before the pattern.}

\item{after}{Integer. Number of words to extract after the pattern.}
}
\value{
A list of windows around each occurrences of the focal pattern.
}
\description{
Extract windows of words around a regex pattern.
}
\author{
Nicolas Mangin
}
