% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-detect-data.R
\name{filter_detect_data}
\alias{filter_detect_data}
\title{Filter Detect Data}
\usage{
filter_detect_data(data, capture = data$capture)
}
\arguments{
\item{data}{The lex_data object.}

\item{capture}{A data frame of the capture data to use.}
}
\value{
A lex_data object.
}
\description{
Filter Detect Data
}

