% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{lexr}
\alias{lexr}
\alias{lexr-package}
\title{Lake Exploitation Study Data Plotting and Analysis}
\description{
Lake Exploitation Study Data Plotting and Analysis
}
\examples{
\dontrun{
library(dplyr) # so tbl data frames print nice
library(klexdatr) # data set for example
library(lexr)

qlex <- input_lex_data("qlexdatr")
check_lex_data(qlex)
print(qlex)
plot(qlex, all = TRUE)

qdetect <- make_detect_data(qlex)
check_detect_data(qdetect)
print(qdetect)
plot(qdetect)

klex <- input_lex_data("klexdatr")
check_lex_data(klex)
print(klex)
plot(klex, all = TRUE)

kdetect <- make_detect_data(klex)
check_detect_data(kdetect)
print(kdetect)
plot(kdetect)
}
}

