% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{lexr}
\alias{lexr}
\alias{lexr-package}
\title{Lake Exploitation Study Data Plotting and Analysis}
\description{
\code{lexr} is an R package to plot and analyse lake exploitation data.
}
\details{
For an example run the demo below.
}
\examples{
\dontrun{
demo(klex, ask = FALSE)
}
}

