% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{lexr}
\alias{lexr}
\alias{lexr-package}
\title{Exploitation Study Analysis}
\description{
\code{lexr} is the R data package.
}
\examples{
\donttest{
pdf("klex.pdf")

plot_section()

dev.off()
}
}

