% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{lexr}
\alias{lexr}
\alias{lexr-package}
\title{Lake Exploitation Study Data Plotting and Analysis}
\description{
Lake Exploitation Study Data Plotting and Analysis
}
\examples{
library(dplyr)
library(magrittr)
library(qlexdatr)
library(lexr)
lex_data <- input_lex_data("qlexdatr")
lex_data \%<>\% check_lex_data()
print(lex_data)
plot(lex_data)
}

