% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{lexr}
\alias{lexr}
\alias{lexr-package}
\title{Lake Exploitation Study Data Plotting and Analysis}
\description{
Lake Exploitation Study Data Plotting and Analysis
}
\examples{
\dontrun{
demo(klex, ask = FALSE)
demo(qlex, ask = FALSE)
}
}

