% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{lexr}
\alias{lexr}
\alias{lexr-package}
\title{Lake Exploitation Study Data Plotting and Analysis}
\description{
Lake Exploitation Study Data Plotting and Analysis
}
\examples{

library(dplyr) # so tbl data frames print nice
library(magrittr) # for piping operator
library(qlexdatr) # data set for example
library(lexr)
lex <- input_lex_data("qlexdatr")
check_lex_data(lex)
print(lex)
plot(lex)
detect <- make_detect_data(lex)
# check_detect_data(detect)
print(detect)
plot(detect)
}

