% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-analysis-data.R
\name{make_analysis_data}
\alias{make_analysis_data}
\title{Make Analysis Data}
\usage{
make_analysis_data(data, capture = data$capture, section = data$section,
  interval_period = get_difftime(data), growth = growth_no,
  spawning = NULL, ...)
}
\arguments{
\item{data}{A detect_data object to use.}

\item{capture}{A data frame of the capture data to use.}

\item{section}{A data frame of the section data to use.}

\item{interval_period}{A difftime element that will be used to group the interval or
a vector indicating the actual interval groupings.}

\item{growth}{A function that takes the length of a fish at capture and predicts
its length after a number of years.}

\item{additional}{arguments passed to growth.}
}
\value{
A detect_data object.
}
\description{
Makes analysis_data object from a detect_data object.
capture$PeriodCapture indicates the period during which the fish was caught.
capture$PeriodTagExpire indicates the period during which the tag expired.
detected is a logical matrix indicating for each individual-period whether it
was detected during the period.
moved is a logical matrix indicating for each individual-period whether it
was detected to have moved during the period
(based on being detected at multiple sections).
reported is a logical matrix indicating for each individual-period
whether it was reported to have been recaught during the period.
released is a logical matrix indicating for each individual-period whether it
was released during the period.
tags is a logical array indicating for each individual-period-tbartag whether
it was attached at the start of the period (note considered attached at the
start of the period during which first caught).
}
\details{
If a difftime element, interval_period cannot be greater than 28 days
i.e. \code{lubridate::make_difftime(60 * 60 * 24 * 28)}.
}

