% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_station}
\alias{plot_station}
\title{Plot Section}
\usage{
plot_station(station, section_polygons = NULL)
}
\arguments{
\item{station}{A data frame of the section data.}

\item{section_polygons}{A xx.}
}
\value{
A ggplot2 object.
}
\description{
Plot Section
}
\examples{
plot_station(qlexdatr::section, qlexdatr::section_polygons)
}

