% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spawning.R
\name{spawning_no}
\alias{spawning_no}
\title{No Spawning}
\usage{
spawning_no(detection, period)
}
\arguments{
\item{detection}{A data.frame of the detection data for the capture.}

\item{period}{A data.frame of the periods.}
}
\description{
A function that simply returns FALSE for every time period.
}
\details{
To identify spawning events when making analysis data pass a custom
function in place of spawning_no. The function has to take the same
arguments and return a logical vector even if there are no detections
for an individual. It should do this by returning FALSE for all
periods outside the spawning window and NA for all periods inside
the spawning window if no information is available.

To see the columns and types in detection and period view
the function definition for \code{spawning_no}.
}

